+++ lib/compiler/include/compiler/syntax/builtins.hh
@@ -24,6 +24,8 @@ constexpr builtin_t INT_CAST{"@intCast", token_type_t::BUILTIN_INT_CAST};
 constexpr builtin_t TRUNCATE{"@truncate", token_type_t::BUILTIN_TRUNCATE};
 constexpr builtin_t BOOL_FROM_INT{"@boolFromInt", token_type_t::BUILTIN_BOOL_FROM_INT};
 constexpr builtin_t INT_FROM_BOOL{"@intFromBool", token_type_t::BUILTIN_INT_FROM_BOOL};
+constexpr builtin_t BACKING_INT{"@backingInt", token_type_t::BUILTIN_BACKING_INT};
+constexpr builtin_t FROM_BACKING_INT{"@fromBackingInt", token_type_t::BUILTIN_FROM_BACKING_INT};
 constexpr builtin_t DYN_CAST{"@dynCast", token_type_t::BUILTIN_DYN_CAST};
 
 constexpr builtin_t INT_FROM_PTR{"@intFromPtr", token_type_t::BUILTIN_INT_FROM_PTR};
@@ -148,6 +150,8 @@ constexpr std::array ALL_BUILTINS{
     builtins::TRUNCATE,
     builtins::BOOL_FROM_INT,
     builtins::INT_FROM_BOOL,
+    builtins::BACKING_INT,
+    builtins::FROM_BACKING_INT,
     builtins::INT_FROM_PTR,
     builtins::PTR_FROM_INT,
     builtins::PTR_FROM_ARRAY,